refactor(frontend): rename ETH send button
